+++ src/useControlled.ts
@@ -1,11 +1,6 @@
 import { useEffect, useRef } from "react";
 
-import {
-  ControlledConfig,
-  ControlledReturn,
-  FieldProps,
-  FormValues,
-} from "./types";
+import { ControlledConfig, ControlledReturn, FormValues } from "./types";
 import * as shared from "./shared";
 import { get, invariant, isFieldElement, isUndefined, warn } from "./utils";
 import useFormState from "./useFormState";
@@ -24,14 +19,14 @@ export default <V extends FormValues = FormValues>(
 ): ControlledReturn => {
   invariant(
     !name,
-    '💡 react-cool-form > useControlled: Missing the "name" parameter.'
+    '💡 react-cool-form > useControlled: Missing "name" parameter.'
   );
 
   const methods = shared.get(formId);
 
   invariant(
     !methods,
-    '💡 react-cool-form > useControlled: You must provide the corresponding ID to the "useForm" hook. See: https://react-cool-form.netlify.app/docs/api-reference/use-form#id'
+    '💡 react-cool-form > useControlled: You must provide the corresponding ID to "useForm" hook. See: https://react-cool-form.netlify.app/docs/api-reference/use-form#id'
   );
 
   const hasWarn = useRef(false);
@@ -48,7 +43,7 @@ export default <V extends FormValues = FormValues>(
     shouldRemoveField,
     defaultValuesRef,
     initialStateRef,
-    controllersRef,
+    controlledsRef,
     fieldValidatorsRef,
     changedFieldRef,
     getNodeValue,
@@ -65,7 +60,7 @@ export default <V extends FormValues = FormValues>(
     [name, removeField, shouldRemoveField]
   );
 
-  controllersRef.current[name] = true;
+  controlledsRef.current[name] = true;
   if (validate) fieldValidatorsRef.current[name] = validate;
 
   let value;
@@ -78,7 +73,7 @@ export default <V extends FormValues = FormValues>(
       setDefaultValue(name, value);
     } else if (!hasWarn.current) {
       warn(
-        `💡 react-cool-form > useControlled: Please provide a default value for the "${name}" field.`
+        `💡 react-cool-form > useControlled: Please provide a default value for "${name}" field.`
       );
       hasWarn.current = true;
     }
@@ -88,36 +83,34 @@ export default <V extends FormValues = FormValues>(
   value = (format ? format(value) : value) ?? "";
   const { onChange, onBlur, ...restProps } = props;
 
-  const fieldProps: FieldProps = {
-    name,
-    value,
-    onChange: (...event) => {
-      let val;
-
-      if (parse) {
-        val = parse(...event);
-      } else {
-        const e = event[0];
-        val =
-          e?.nativeEvent instanceof Event && isFieldElement(e.target)
-            ? getNodeValue(name)
-            : e;
-      }
-
-      handleChangeEvent(name, val);
-      if (onChange) onChange(...event);
-      changedFieldRef.current = name;
-    },
-    onBlur: (e) => {
-      setTouchedMaybeValidate(name);
-      if (onBlur) onBlur(e);
-      changedFieldRef.current = undefined;
-    },
-    ...restProps,
-  };
-
   return [
-    fieldProps,
+    {
+      name,
+      value,
+      onChange: (...event) => {
+        let val;
+
+        if (parse) {
+          val = parse(...event);
+        } else {
+          const e = event[0];
+          val =
+            e?.nativeEvent instanceof Event && isFieldElement(e.target)
+              ? getNodeValue(name)
+              : e;
+        }
+
+        handleChangeEvent(name, val);
+        if (onChange) onChange(...event);
+        changedFieldRef.current = name;
+      },
+      onBlur: (e) => {
+        setTouchedMaybeValidate(name);
+        if (onBlur) onBlur(e);
+        changedFieldRef.current = undefined;
+      },
+      ...restProps,
+    },
     { error: meta.error, isTouched: !!meta.isTouched, isDirty: !!meta.isDirty },
   ];
 };